dom/tests/mochitest/geolocation/test_clearWatchBeforeAllowing.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=886026
-->
<head>
  <title>Test for getCurrentPosition </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="geolocation_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank"
href="https://bugzilla.mozilla.org/show_bug.cgi?id=886026">Mozilla Bug 886026</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">
"use strict";

SimpleTest.waitForExplicitFinish();
resume_geolocationProvider(function() {
  force_prompt(true, run_test);
});

function run_test() {
  var successCallbackCalled = false,
  errorCallbackCalled = false;

  var watchId = navigator.geolocation.watchPosition(
    function(pos) {
      check_geolocation(pos);
      successCallbackCalled = true;
    }, function(err) {
      errorCallbackCalled = true;
    }
  );

  navigator.geolocation.getCurrentPosition(
    function(pos) {
      check_geolocation(pos);
      SimpleTest.executeSoon(function() {
        ok(successCallbackCalled == false,
        "getCurrentPosition : Success callback should not have been called");

        ok(errorCallbackCalled == false,
        "getCurrentPosition : Error callback should not have been called");

        SimpleTest.finish();
      });
    }
  );

  navigator.geolocation.clearWatch(watchId);
}
</script>
</pre>
</body>
</html>
